requirements.txt
wheel
pycairo
shapely

numpy
opencv-python

pylint
autopep8

-r lib/requirements.txt
